requirements.txt
certifi==2020.6.20
grpcio==1.30.0
grpcio-tools==1.30.0
protobuf==3.12.4
six==1.15.0
wincertstore==0.2
